+++ examples/take_picture_raw/src/app_raw.h
@@ -3,9 +3,8 @@
 
 #pragma once
 
-#include "xs1.h"
-#include "platform.h"
-#include "xccompat.h"
+#include <xs1.h>
+#include <platform.h>
 
 #include "camera_main.h"