+++ python/decode_downsampled.py
@@ -1,4 +1,4 @@
-# Copyright 2023 XMOS LIMITED.
+# Copyright 2023-2024 XMOS LIMITED.
 # This Software is subject to the terms of the XMOS Public Licence: Version 1.
 
 """
@@ -11,27 +11,46 @@ import cv2
 import matplotlib.pyplot as plt
 import numpy as np
 from PIL import Image  # just to avoid color BGR issues when writting
-from dotenv import load_dotenv
-load_dotenv()  # take environment variables from .env.
+import argparse
+import sys
 
 from utils import show_histogram_by_channel
 
-input_name = os.getenv('BINARY_IMG_PATH') or "capture.bin"
 
-# read the data
-with open(input_name, "rb") as f:
-    data = f.read()
+def decode_downsampled_image(input_name="capture.bin", width=160, height=120, img_dtype="uint8", plot=True):
+    # read the data
+    with open(input_name, "rb") as f:
+        data = f.read()
 
-# unpack
-width, height = 160, 120
-buffer = np.frombuffer(data, dtype=np.uint8)
-img = buffer.reshape(height, width, 3)
-print("unpacked_data")
+    # unpack the data
+    if img_dtype == "uint8":
+        buffer = np.frombuffer(data, dtype=np.uint8)
+    elif img_dtype == "int8":
+        buffer = np.frombuffer(data, dtype=np.int8)
+        buffer = buffer + 128
+        buffer = buffer.astype(np.uint8)
+    else:
+        raise ValueError(f"Invalid dtype: {img_dtype}")
+    
+    img = buffer.reshape(height, width, 3)
 
-# show image
-plt.figure()
-plt.imshow(img)
-plt.show()
+    if plot:
+        # show image
+        plt.figure()
+        plt.imshow(img)
+        plt.show()
 
-# show histograms
-show_histogram_by_channel(img, 3000)
+        # show histograms
+        show_histogram_by_channel(img, 3000)
+    return img
+
+
+if __name__ == "__main__":
+    argparse = argparse.ArgumentParser()
+    argparse.add_argument("--input", help="input file name", default="capture.bin")
+    argparse.add_argument("--width", help="image width", default=160, type=int)
+    argparse.add_argument("--height", help="image height", default=120, type=int)
+    argparse.add_argument("--dtype", help="image dtype", default="uint8")
+    args = argparse.parse_args()
+    decode_downsampled_image(args.input, args.width, args.height, args.dtype)
+    print("done")