+++ CHANGELOG.rst
@@ -1,5 +1,13 @@
-fwk_camera change log
+lib_camera change log
 =====================
+
+1.0.0
+-----
+
+  * ADDED: functions for crop and resize. See "isp_functions.h"
+  * ADDED: Support for the XCORE-VISION-EXPLORER board
+  * ADDED: Support for Xcommon_cmake
+  * CHANGED: Repository structure
 
 0.2.1
 -----